Adds alt attributes to images for country flags and relay flags (Fixes: #21291)
In cases where there is no other text and the meaning is conveyed only in the
icon, this adds alt text explaining the icon. In cases where the icon is next
to text that explains the icon, the alt attribute is specified as an empty
string as recommended by the HTML 5 specification.