Removes graphs that no longer have Onionoo data to back them (Fixes: #19553)
This removes the 3 days and 1 week bandwidth charts and also removes references
to the 3 days clients chart as the Onionoo specification doesn't claim to
provide this chart either.

In the future we may attempt to produce these graphs from the data provided by
Onionoo, but this will require more than simply putting the right data on the
screen and I'd like to keep Atlas as thin a layer as possible on top of
Onionoo.