Removes duplicate id attributes on SVG graphs (Fixes: #21351)
Comment from cypherpunks: "A quick grep shows there is no CSS that uses the
bw_month id. The id is only used by the JS code that generates the SVG graphs
to know where to inject the graphs. This code also injects the duplicate id for
some unknown reason. The inline patch fixes the issue by simply removing the
injection of the duplicate id."

I would agree with this comment.