Uses pre elements for fingerprints in details view, not tt (Fixes: #21350)
Comment from cypherpunks: "It reverts the solution from #12685 because <pre>
tags wrap their content automatically. However, using the <pre> tag changes the
fingerprint to look like the fingerprints of the family members (except with
black text instead of blue or orange). IMO this is preferable because it makes
displaying fingerprints consistent but thought I'd mention it anyway."

In testing, I found this to actually be a better solution for responsiveness as
instead of wrapping at a fixed location, everything can still be viewable even
with a 160px wide screen. This does lead to having different length lines when
wrapping, and I guess user feedback will let us know if this was a dreadful
idea, but I don't see it causing disasterous breakage for anyone.