Merge the current status and properties columns
Moving the relay and bridge status next to its name made the current
status column contain a single item (the uptime or downtime). Because
this is a waste of screen space the current status column is merged into
the properties column.

The merge includes some reordering to ensure the fingerprint of the
relay and bridge is always the first item.