Support bandwidth values that are zero
The boolean value of zero is false in JavaScript. This means that
bandwidth values that are zero are set to 'null'. Null values result in
empty fields in the router details and search pages.

The checks are meant to catch undefined variables so they are replaced
by typeof comparisons which are stricter.

Closes #22039.