Hash the fingerprints from summary documents
The summary documents are used for searches and form the basis for
subsequent details document requests. The details documents requests
use the lookup parameter which supports hashed and double hashed
fingerprints to prevent users leaking bridge fingerprints.

If a relay and a bridge share the same fingerprint FP (for example when
a bridge becomes a relay without changing its fingerprint) the summary
document for fingerprint FP contains both the relay fingerprint RP
(where RP == FP) and the hashed bridge fingerprint BP (where BP ==
H(FP)).

The details document request for RP returns only the relay because RP ==
FP and the details document request for BP returns the relay and the
bridge because BP == H(FP) == H(RP). Relays take preference over bridges
when both are returned and results in Atlas displaying the relay details
twice on the search page.

Hashing the fingerprints in the summary document results in details
document requests for BP and H(BP) where BP returns the relay and bridge
as before but where H(BP) returns only the bridge.

Closes #21615.